requirements.txt
yt-dlp
pillow
imagehash
